project/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
python-dotenv==1.0.0
retry==0.9.2